♻️ refactor: refactor minmaxlttb files
