--- downsample_rs/src/minmaxlttb/mod.rs
@@ -1,5 +1,0 @@
-pub mod scalar;
-pub use scalar::*;
-pub mod simd;
-pub use simd::*;
-mod generic; // Private module
